fix: Fix an edge case with device invitation
This covers an edge case when claiming device invitation, when the profile is there but `useProfile` has not updated yet.
The 'useProfile' does update properly but the component issues a redirection already and has no chance of reacting to that.
Using latch() around 'client.echo.joinHalo' did not help in this regard